ref: remove `utils/exploreFilterCount`
